# app/openapi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Team Task Manager API",
    "description": "Comprehensive API for team task management with authentication, analytics, and file attachments",
    "version": "2.0.0",
    "contact": {
      "name": "Team Task Manager",
      "email": "support@teamtaskmanager.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["PLANNING", "ACTIVE", "ON_HOLD", "COMPLETED", "CANCELLED"] },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "teamId": { "type": "string" },
          "createdById": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "List projects",
        "parameters": [
          {
            "name": "teamId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["PLANNING", "ACTIVE", "ON_HOLD", "COMPLETED", "CANCELLED"] }
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Project" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "teamId"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["PLANNING", "ACTIVE", "ON_HOLD", "COMPLETED", "CANCELLED"], "default": "PLANNING" },
                  "startDate": { "type": "string", "format": "date-time" },
                  "endDate": { "type": "string", "format": "date-time" },
                  "teamId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      }
    },
    "/analytics/tasks": {
      "get": {
        "summary": "Get task analytics",
        "parameters": [
          {
            "name": "teamId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "overview": {
                      "type": "object",
                      "properties": {
                        "totalTasks": { "type": "integer" },
                        "completedTasks": { "type": "integer" },
                        "inProgressTasks": { "type": "integer" },
                        "overdueTasks": { "type": "integer" },
                        "completionRate": { "type": "number" },
                        "averageCompletionTime": { "type": "number" }
                      }
                    },
                    "tasksByStatus": {
                      "type": "object",
                      "properties": {
                        "todo": { "type": "integer" },
                        "inProgress": { "type": "integer" },
                        "review": { "type": "integer" },
                        "done": { "type": "integer" }
                      }
                    },
                    "tasksByPriority": {
                      "type": "object",
                      "properties": {
                        "low": { "type": "integer" },
                        "medium": { "type": "integer" },
                        "high": { "type": "integer" },
                        "urgent": { "type": "integer" }
                      }
                    },
                    "weeklyProgress": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "week": { "type": "string" },
                          "completed": { "type": "integer" },
                          "created": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search across entities",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "minLength": 1 }
          },
          {
            "name": "entityTypes",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Comma-separated list of entity types (TASK,PROJECT,COMMENT)"
          },
          {
            "name": "teamId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "entityType": { "type": "string", "enum": ["TASK", "PROJECT", "COMMENT"] },
                          "title": { "type": "string" },
                          "description": { "type": "string" },
                          "relevanceScore": { "type": "number" }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "hasMore": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}